.claude/worktrees/stoic-hamilton/requirements.txt
python-telegram-bot==21.9
anthropic==0.52.0
aiohttp==3.11.11
pydantic==2.10.4
aiosqlite==0.20.0
ruamel.yaml==0.18.6
structlog==24.4.0
orjson==3.10.13
plotly==5.24.1
kaleido==0.2.1
